+++ src/shared/alerts/index.js
@@ -1,8 +1,8 @@
-import { Alert } from 'react-native';
+import { Alert, ToastAndroid } from 'react-native';
 import i18n from '../i18n';
 
 export function alertDelete({
-  title, message, deleteMessage, onDelete,
+  title, message, deleteMessage, toastMessage, onDelete,
 }) {
   Alert.alert(
     title || i18n.t('deletion.deleteQuestion'),
@@ -11,7 +11,16 @@ export function alertDelete({
       { text: i18n.t('cancel'), style: 'cancel' },
       {
         text: deleteMessage || i18n.t('deletion.delete'),
-        onPress: onDelete,
+        onPress: () => {
+          onDelete();
+          ToastAndroid.showWithGravityAndOffset(
+            toastMessage || i18n.t('deletion.deleted'),
+            ToastAndroid.LONG,
+            ToastAndroid.BOTTOM,
+            0,
+            50,
+          );
+        },
         style: 'destructive', // NOTE: style works only for iOS
       },
     ],